Add ability to apply elements to current system.
This commit adds the ability to apply diskimage-builder style elements
to the current system, as opposed to a chroot.  The functionality is
implemented in new python modules under diskimage_builder.  A script
using this python code at bin/element-runner is also included.  There
are some simple test elements under diskimage_builder/tests/elements to
aid with testing/verification as well.

Change-Id: Ic2bcb8b9f17e03709048ee4702270ede9a95d064